refactors WidgetIcon to be nested inside the modal. Changes the order of the widgets
