+++ client/src/components/SelectWidgetsModal/WidgetIcon/index.js
