+++ client/src/components/SelectWidgetsModal/WidgetIcon/style.module.scss
