+++ client/src/components/SelectWidgetsModal/index.js
@@ -9,7 +9,7 @@ import toDoList from "../../images/toDoList.svg";
 import calendar from "../../images/calendar.svg";
 import timer from "../../images/timer.svg";
 import styles from "./style.module.scss";
-import WidgetIcon from "../WidgetIcon";
+import WidgetIcon from "./WidgetIcon";
 
 const useStyles = makeStyles({
     button: {
